requirements.txt
streamlit
supabase
openai>=1.0.0
python-dotenv